requirements.txt
# pip install -U pip
# pip install --default-timeout=1000 --editable ./fairseq
fastBPE==0.1.0
sacremoses==0.0.47
bitarray==2.5.1
hydra-core==1.0.7
omegaconf==2.0.6
regex==2021.11.10
requests-oauthlib==1.3.0
subword-nmt==0.3.8
torchaudio==0.12.0
omegaconf==2.0.6
SoundFile==0.10.3.post1
moviepy==1.0.3
mysmallutils==1.1.2
